client: add nbd-clid daemon support
