+++ client/gluster.c
@@ -17,90 +17,40 @@
 #include "rpc_nbd.h"
 #include "utils/utils.h"
 #include "nbd-log.h"
-#include "nbd-cli-cmd.h"
-
-struct cli_cmd gluster_cmds[];
-
-static int gluster_help_routine(int count, char **options)
-{
-    int i;
-
-    nbd_info("Usage: \n\n");
-    for (i = 0; gluster_cmds[i].pattern; i++) {
-        if (gluster_cmds[i].disable)
-            continue;
-
-        nbd_info("\t%s\n", gluster_cmds[i].pattern);
-        nbd_info("\t\t%s\n\n", gluster_cmds[i].desc);
-    }
-    nbd_info("\n");
-
-    return 0;
-}
-
-static int gluster_create_routine(int count, char **options)
-{
-    return nbd_create_backstore(count, options, NBD_BACKSTORE_GLUSTER);
-}
-
-static int gluster_delete_routine(int count, char **options)
-{
-    return nbd_delete_backstore(count, options, NBD_BACKSTORE_GLUSTER);
-}
-
-static int gluster_map_routine(int count, char **options)
-{
-    return nbd_map_device(count, options, NBD_BACKSTORE_GLUSTER);
-}
-
-static int gluster_unmap_routine(int count, char **options)
-{
-    return nbd_unmap_device(count, options, NBD_BACKSTORE_GLUSTER);
-}
-
-static int gluster_list_routine(int count, char **options)
-{
-    return nbd_list_devices(count, options, NBD_BACKSTORE_GLUSTER);
-}
-
+#include "nbd-cli-common.h"
 
 struct cli_cmd gluster_cmds[] = {
-    {.pattern = "gluster",
-     .call    = gluster_help_routine,
-     .desc    = "Display help for gluster commands",
-     .disable = true,
-    },
-    {.pattern = "gluster help",
-     .call    = gluster_help_routine,
+    {.pattern = "help",
+     .cmd     = NBD_CLI_HELP,
      .desc    = "Display help for gluster commands",
     },
-    {.pattern = "gluster create <VOLUME/FILEPATH> [prealloc] <size SIZE> [host RUNNER_HOST]",
-     .call    = gluster_create_routine,
+    {.pattern = "create <VOLUME/FILEPATH> [prealloc] <size SIZE> [host RUNNER_HOST]",
+     .cmd     = NBD_CLI_CREATE,
      .desc    = "Create FILEPATH in the VOLUME, prealloc is false as default, and the SIZE is valid\n\t\twith B, K(iB), M(iB), G(iB), T(iB), P(iB), E(iB), Z(iB), Y(iB), RUNNER_HOST will\n\t\tbe 'localhost' as default",
     },
-    {.pattern = "gluster delete <VOLUME/FILEPATH> [host RUNNER_HOST]",
-     .call    = gluster_delete_routine,
+    {.pattern = "delete <VOLUME/FILEPATH> [host RUNNER_HOST]",
+     .cmd     = NBD_CLI_DELETE,
      .desc    = "Delete FILEPATH from the VOLUME, RUNNER_HOST will be 'localhost' as default",
     },
-    {.pattern = "gluster map <VOLUME/FILEPATH> [nbd-device] [timeout TIME] [readonly] [host RUNNER_HOST]",
-     .call    = gluster_map_routine,
+    {.pattern = "map <VOLUME/FILEPATH> [nbd-device] [timeout TIME] [readonly] [host RUNNER_HOST]",
+     .cmd     = NBD_CLI_MAP,
      .desc    = "Map FILEPATH to the nbd device, as default the socket connection timeout is 30 seconds,\n\t\tnone readonly, RUNNER_HOST will be 'localhost' as default",
     },
-    {.pattern = "gluster unmap <nbd-device|VOLUME/FILEPATH> [host RUNNER_HOST]",
-     .call    = gluster_unmap_routine,
+    {.pattern = "unmap <nbd-device|VOLUME/FILEPATH> [host RUNNER_HOST]",
+     .cmd     = NBD_CLI_UNMAP,
      .desc    = "Unmap the nbd device or VOLUME/FILEPATH, RUNNER_HOST will be 'localhost' as default",
     },
-    {.pattern = "gluster list [map|unmap|create|dead|live|all] [host RUNNER_HOST]",
-     .call    = gluster_list_routine,
+    {.pattern = "list [map|unmap|create|dead|live|all] [host RUNNER_HOST]",
+     .cmd     = NBD_CLI_LIST,
      .desc    = "List the mapped|unmapped NBD devices or the created|dead|live backstores, all as\n\t\tdefault. 'create' means the backstores are just created or unmapped. 'dead' means\n\t\tthe IO connection is lost, this is mainly due to the nbd-runner service is restart\n\t\twithout unmapping. 'live' means everything is okay for both mapped and IO connection,\n\t\tRUNNER_HOST will be 'localhost' as default"
     },
     {.pattern = NULL,
-     .call    = NULL,
+     .cmd     = NBD_CLI_MAX,
      .desc    = NULL,
     },
 };
 
-int cli_cmd_gluster_register(GHashTable *cmds_hash)
+int cli_cmd_gluster_register(GPtrArray *cmds_list, cmd_fn_t fn)
 {
-    return nbd_register_cmds(cmds_hash, gluster_cmds);
+    return fn(cmds_list, gluster_cmds);
 }
