+++ configure.ac
@@ -153,6 +153,7 @@ AC_CONFIG_HEADERS([config.h])
 
 AC_CONFIG_FILES([Makefile
                  nbd-runner.service
+                 nbd-clid.service
                  rpc/Makefile
                  client/Makefile
                  server/Makefile
